# cli/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"HttpExceptionBodyMessage": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					{
						"type": "number",
						"format": "double"
					}
				]
			},
			"HttpExceptionBody": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/HttpExceptionBodyMessage"
					},
					"error": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ContractType": {
				"enum": [
					"AccountCreateContract",
					"TransferContract",
					"TransferAssetContract",
					"VoteAssetContract",
					"VoteWitnessContract",
					"WitnessCreateContract",
					"AssetIssueContract",
					"WitnessUpdateContract",
					"ParticipateAssetIssueContract",
					"AccountUpdateContract",
					"FreezeBalanceContract",
					"UnfreezeBalanceContract",
					"CancelAllUnfreezeV2Contract",
					"WithdrawBalanceContract",
					"UnfreezeAssetContract",
					"UpdateAssetContract",
					"ProposalCreateContract",
					"ProposalApproveContract",
					"ProposalDeleteContract",
					"SetAccountIdContract",
					"CustomContract",
					"CreateSmartContract",
					"TriggerSmartContract",
					"GetContract",
					"UpdateSettingContract",
					"ExchangeCreateContract",
					"ExchangeInjectContract",
					"ExchangeWithdrawContract",
					"ExchangeTransactionContract",
					"UpdateEnergyLimitContract",
					"AccountPermissionUpdateContract",
					"ClearABIContract",
					"UpdateBrokerageContract",
					"ShieldedTransferContract",
					"MarketSellAssetContract",
					"MarketCancelOrderContract",
					"FreezeBalanceV2Contract",
					"UnfreezeBalanceV2Contract",
					"WithdrawExpireUnfreezeContract",
					"DelegateResourceContract",
					"UnDelegateResourceContract",
					"UNRECOGNIZED"
				],
				"type": "string"
			},
...
			"ContractParamterWrapper_ContractParamter_": {
				"properties": {
					"value": {
						"$ref": "#/components/schemas/ContractParamter"
					},
					"type_url": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"type_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionContract_ContractParamter_": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/ContractType"
					},
					"parameter": {
						"$ref": "#/components/schemas/ContractParamterWrapper_ContractParamter_"
					},
					"Permission_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type",
					"parameter"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Transaction": {
				"properties": {
					"visible": {
						"type": "boolean"
					},
					"txID": {
						"type": "string"
					},
					"raw_data": {
						"properties": {
							"fee_limit": {},
							"data": {},
							"timestamp": {
								"type": "number",
								"format": "double"
							},
							"expiration": {
								"type": "number",
								"format": "double"
							},
							"ref_block_hash": {
								"type": "string"
							},
							"ref_block_bytes": {
								"type": "string"
							},
							"contract": {
								"items": {
									"$ref": "#/components/schemas/TransactionContract_ContractParamter_"
								},
								"type": "array"
							}
						},
						"required": [
							"timestamp",
							"expiration",
							"ref_block_hash",
							"ref_block_bytes",
							"contract"
						],
						"type": "object"
					},
					"raw_data_hex": {
						"type": "string"
					}
				},
				"required": [
					"visible",
					"txID",
					"raw_data",
					"raw_data_hex"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RawTronTransaction": {
				"$ref": "#/components/schemas/Transaction"
			},
			"EssentialWeb3Transaction": {
				"description": "The `EssentialWeb3Transaction` interface provides the minimum set of Web3 Transaction\nreturned by the SDK. It allows any web3-like provider to be used, as long as it\nmatches the signature of these essential interface. This prevents\ntight coupling to a specific version of the `web3` library.",
				"properties": {
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"RawEvmTransaction": {
				"$ref": "#/components/schemas/EssentialWeb3Transaction"
			},
			"RawSorobanTransaction": {
				"type": "string"
			},
			"RawAlgTransaction": {
				"items": {
					"type": "string"
				},
				"type": "array"
			},
			"RawStxTransaction": {
				"type": "string"
			},
			"RawSuiTransaction": {
				"type": "string"
			},
			"Uint8Array": {
				"description": "A typed array of 8-bit unsigned integer values. The contents are initialized to 0. If the\nrequested number of bytes could not be allocated an exception is raised.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"MessageHeader": {
				"properties": {
					"numReadonlyUnsignedAccounts": {
						"type": "number",
						"format": "double",
						"description": "The last `numReadonlySignedAccounts` of the unsigned keys are read-only accounts"
					},
					"numReadonlySignedAccounts": {
						"type": "number",
						"format": "double",
						"description": "The last `numReadonlySignedAccounts` of the signed keys are read-only accounts"
					},
					"numRequiredSignatures": {
						"type": "number",
						"format": "double",
						"description": "The number of signatures required for this message to be considered valid. The\nsignatures must match the first `numRequiredSignatures` of `accountKeys`."
					}
				},
				"required": [
					"numReadonlyUnsignedAccounts",
					"numReadonlySignedAccounts",
					"numRequiredSignatures"
				],
				"type": "object",
				"description": "The message header, identifying signed and read-only account"
			},
			"PublicKey": {
				"type": "string",
				"description": "A public key"
			},
			"Blockhash": {
				"type": "string",
				"description": "Blockhash as Base58 string."
			},
			"CompiledInstruction": {
				"properties": {
					"data": {
						"type": "string",
						"description": "The program input data encoded as base 58"
					},
					"accounts": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "Ordered indices into the transaction keys array indicating which accounts to pass to the program"
					},
					"programIdIndex": {
						"type": "number",
						"format": "double",
						"description": "Index into the transaction keys array indicating the program account that executes this instruction"
					}
				},
				"required": [
					"data",
					"accounts",
					"programIdIndex"
				],
				"type": "object",
				"description": "An instruction to execute by a program"
			},
			"Message": {
				"description": "List of instructions to be processed atomically",
				"properties": {
					"header": {
						"$ref": "#/components/schemas/MessageHeader"
					},
					"accountKeys": {
						"items": {
							"$ref": "#/components/schemas/PublicKey"
						},
						"type": "array"
					},
					"recentBlockhash": {
						"$ref": "#/components/schemas/Blockhash"
					},
					"instructions": {
						"items": {
							"$ref": "#/components/schemas/CompiledInstruction"
						},
						"type": "array"
					}
				},
				"required": [
					"header",
					"accountKeys",
					"recentBlockhash",
					"instructions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MessageV0": {
				"properties": {
					"header": {
						"$ref": "#/components/schemas/MessageHeader"
					},
					"staticAccountKeys": {
						"items": {
							"$ref": "#/components/schemas/PublicKey"
						},
						"type": "array"
					},
					"recentBlockhash": {
						"$ref": "#/components/schemas/Blockhash"
					},
					"compiledInstructions": {
						"items": {
							"$ref": "#/components/schemas/MessageCompiledInstruction"
						},
						"type": "array"
					},
					"addressTableLookups": {
						"items": {
							"$ref": "#/components/schemas/MessageAddressTableLookup"
						},
						"type": "array"
					}
				},
				"required": [
					"header",
					"staticAccountKeys",
					"recentBlockhash",
					"compiledInstructions",
					"addressTableLookups"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VersionedMessage": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Message"
					},
					{
						"$ref": "#/components/schemas/MessageV0"
					}
				]
			},
			"VersionedTransaction": {
				"description": "Versioned transaction class",
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Uint8Array"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/VersionedMessage"
					}
				},
				"required": [
					"signatures",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RawBridgeSolanaTransaction": {
				"$ref": "#/components/schemas/VersionedTransaction"
			},
			"SignaturePubkeyPair": {
				"properties": {
					"publicKey": {
						"$ref": "#/components/schemas/PublicKey"
					},
					"signature": {
						"type": "string",
						"format": "byte",
						"nullable": true
					}
				},
				"required": [
					"publicKey",
					"signature"
				],
				"type": "object",
				"description": "Pair of signature and corresponding public key"
			},
			"AccountMeta": {
				"properties": {
					"isWritable": {
						"type": "boolean",
						"description": "True if the `pubkey` can be loaded as a read-write account."
					},
					"isSigner": {
						"type": "boolean",
						"description": "True if an instruction requires a transaction signature matching `pubkey`"
					},
					"pubkey": {
						"$ref": "#/components/schemas/PublicKey",
						"description": "An account's public key"
					}
				},
				"required": [
					"isWritable",
					"isSigner",
					"pubkey"
				],
				"type": "object",
				"description": "Account metadata used to define instructions"
			},
			"TransactionInstruction": {
				"description": "Transaction Instruction class",
				"properties": {
					"keys": {
						"items": {
							"$ref": "#/components/schemas/AccountMeta"
						},
						"type": "array",
						"description": "Public keys to include in this transaction\nBoolean represents whether this pubkey needs to sign the transaction"
					},
					"programId": {
						"$ref": "#/components/schemas/PublicKey",
						"description": "Program Id to execute"
					},
					"data": {
						"type": "string",
						"format": "byte",
						"description": "Program input"
					}
				},
				"required": [
					"keys",
					"programId",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NonceInformation": {
				"properties": {
					"nonceInstruction": {
						"$ref": "#/components/schemas/TransactionInstruction",
						"description": "AdvanceNonceAccount Instruction"
					},
					"nonce": {
						"$ref": "#/components/schemas/Blockhash",
						"description": "The current blockhash stored in the nonce"
					}
				},
				"required": [
					"nonceInstruction",
					"nonce"
				],
				"type": "object",
				"description": "Nonce information to be used to build an offline Transaction."
			},
			"SolanaWeb3Transaction": {
				"description": "Transaction class",
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/SignaturePubkeyPair"
						},
						"type": "array",
						"description": "Signatures for the transaction.  Typically created by invoking the\n`sign()` method"
					},
					"feePayer": {
						"$ref": "#/components/schemas/PublicKey",
						"description": "The transaction fee payer"
					},
					"instructions": {
						"items": {
							"$ref": "#/components/schemas/TransactionInstruction"
						},
						"type": "array",
						"description": "The instructions to atomically execute"
					},
					"recentBlockhash": {
						"$ref": "#/components/schemas/Blockhash",
						"description": "A recent transaction id. Must be populated by the caller"
					},
					"lastValidBlockHeight": {
						"type": "number",
						"format": "double",
						"description": "the last block chain can advance to before tx is declared expired"
					},
					"nonceInfo": {
						"$ref": "#/components/schemas/NonceInformation",
						"description": "Optional Nonce information. If populated, transaction will use a durable\nNonce hash instead of a recentBlockhash. Must be populated by the caller"
					},
					"minNonceContextSlot": {
						"type": "number",
						"format": "double",
						"description": "If this is a nonce transaction this represents the minimum slot from which\nto evaluate if the nonce has advanced when attempting to confirm the\ntransaction. This protects against a case where the transaction confirmation\nlogic loads the nonce account from an old slot and assumes the mismatch in\nnonce value implies that the nonce has been advanced."
					}
				},
				"required": [
					"signatures",
					"instructions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RawPoolSolanaTransaction": {
				"$ref": "#/components/schemas/SolanaWeb3Transaction"
			},
			"SdkRawTransaction": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/RawTronTransaction"
					},
					{
						"$ref": "#/components/schemas/RawEvmTransaction"
					},
					{
						"$ref": "#/components/schemas/RawSorobanTransaction"
					},
					{
						"$ref": "#/components/schemas/RawAlgTransaction"
					},
					{
						"$ref": "#/components/schemas/RawStxTransaction"
					},
					{
						"$ref": "#/components/schemas/RawSuiTransaction"
					},
					{
						"$ref": "#/components/schemas/RawBridgeSolanaTransaction"
					},
					{
						"$ref": "#/components/schemas/RawPoolSolanaTransaction"
					}
				]
			},
			"RawTransaction": {
				"$ref": "#/components/schemas/SdkRawTransaction"
			},
			"HorizonApi.SubmitTransactionResponse": {
				"properties": {
					"hash": {
						"type": "string"
					},
					"ledger": {
						"type": "number",
						"format": "double"
					},
					"successful": {
						"type": "boolean"
					},
					"envelope_xdr": {
						"type": "string"
					},
					"result_xdr": {
						"type": "string"
					},
					"result_meta_xdr": {
						"type": "string"
					},
					"paging_token": {
						"type": "string"
					}
				},
				"required": [
					"hash",
					"ledger",
					"successful",
					"envelope_xdr",
					"result_xdr",
					"result_meta_xdr",
					"paging_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"xdr.TransactionResult": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"xdr.DiagnosticEvent": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Api.SendTransactionStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"DUPLICATE",
					"TRY_AGAIN_LATER",
					"ERROR"
				]
			},
			"Api.SendTransactionResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Api.SendTransactionStatus"
					},
					"hash": {
						"type": "string"
					},
					"latestLedger": {
						"type": "number",
						"format": "double"
					},
					"latestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"errorResult": {
						"$ref": "#/components/schemas/xdr.TransactionResult"
					},
					"diagnosticEvents": {
						"items": {
							"$ref": "#/components/schemas/xdr.DiagnosticEvent"
						},
						"type": "array"
					}
				},
				"required": [
					"status",
					"hash",
					"latestLedger",
					"latestLedgerCloseTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Api.GetTransactionStatus.SUCCESS": {
				"enum": [
					"SUCCESS"
				],
				"type": "string"
			},
			"xdr.TransactionEnvelope": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"xdr.TransactionMeta": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"xdr.ScVal": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"xdr.TransactionEvent": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"xdr.ContractEvent": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Api.TransactionEvents": {
				"properties": {
					"transactionEventsXdr": {
						"items": {
							"$ref": "#/components/schemas/xdr.TransactionEvent"
						},
						"type": "array"
					},
					"contractEventsXdr": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/xdr.ContractEvent"
							},
							"type": "array"
						},
						"type": "array"
					}
				},
				"required": [
					"transactionEventsXdr",
					"contractEventsXdr"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Api.GetSuccessfulTransactionResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Api.GetTransactionStatus.SUCCESS"
					},
					"txHash": {
						"type": "string"
					},
					"latestLedger": {
						"type": "number",
						"format": "double"
					},
					"latestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"oldestLedger": {
						"type": "number",
						"format": "double"
					},
					"oldestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"ledger": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"applicationOrder": {
						"type": "number",
						"format": "double"
					},
					"feeBump": {
						"type": "boolean"
					},
					"envelopeXdr": {
						"$ref": "#/components/schemas/xdr.TransactionEnvelope"
					},
					"resultXdr": {
						"$ref": "#/components/schemas/xdr.TransactionResult"
					},
					"resultMetaXdr": {
						"$ref": "#/components/schemas/xdr.TransactionMeta"
					},
					"diagnosticEventsXdr": {
						"items": {
							"$ref": "#/components/schemas/xdr.DiagnosticEvent"
						},
						"type": "array"
					},
					"returnValue": {
						"$ref": "#/components/schemas/xdr.ScVal"
					},
					"events": {
						"$ref": "#/components/schemas/Api.TransactionEvents"
					}
				},
				"required": [
					"status",
					"txHash",
					"latestLedger",
					"latestLedgerCloseTime",
					"oldestLedger",
					"oldestLedgerCloseTime",
					"ledger",
					"createdAt",
					"applicationOrder",
					"feeBump",
					"envelopeXdr",
					"resultXdr",
					"resultMetaXdr",
					"events"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Api.GetTransactionStatus.FAILED": {
				"enum": [
					"FAILED"
				],
				"type": "string"
			},
			"Api.GetFailedTransactionResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Api.GetTransactionStatus.FAILED"
					},
					"txHash": {
						"type": "string"
					},
					"latestLedger": {
						"type": "number",
						"format": "double"
					},
					"latestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"oldestLedger": {
						"type": "number",
						"format": "double"
					},
					"oldestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"ledger": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"applicationOrder": {
						"type": "number",
						"format": "double"
					},
					"feeBump": {
						"type": "boolean"
					},
					"envelopeXdr": {
						"$ref": "#/components/schemas/xdr.TransactionEnvelope"
					},
					"resultXdr": {
						"$ref": "#/components/schemas/xdr.TransactionResult"
					},
					"resultMetaXdr": {
						"$ref": "#/components/schemas/xdr.TransactionMeta"
					},
					"diagnosticEventsXdr": {
						"items": {
							"$ref": "#/components/schemas/xdr.DiagnosticEvent"
						},
						"type": "array"
					},
					"events": {
						"$ref": "#/components/schemas/Api.TransactionEvents"
					}
				},
				"required": [
					"status",
					"txHash",
					"latestLedger",
					"latestLedgerCloseTime",
					"oldestLedger",
					"oldestLedgerCloseTime",
					"ledger",
					"createdAt",
					"applicationOrder",
					"feeBump",
					"envelopeXdr",
					"resultXdr",
					"resultMetaXdr",
					"events"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Api.GetTransactionStatus.NOT_FOUND": {
				"enum": [
					"NOT_FOUND"
				],
				"type": "string"
			},
			"Api.GetMissingTransactionResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Api.GetTransactionStatus.NOT_FOUND"
					},
					"txHash": {
						"type": "string"
					},
					"latestLedger": {
						"type": "number",
						"format": "double"
					},
					"latestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					},
					"oldestLedger": {
						"type": "number",
						"format": "double"
					},
					"oldestLedgerCloseTime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"txHash",
					"latestLedger",
					"latestLedgerCloseTime",
					"oldestLedger",
					"oldestLedgerCloseTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Api.GetTransactionResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Api.GetSuccessfulTransactionResponse"
					},
					{
						"$ref": "#/components/schemas/Api.GetFailedTransactionResponse"
					},
					{
						"$ref": "#/components/schemas/Api.GetMissingTransactionResponse"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "3.30.0-alpha.2",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/utils/solana/add-memo": {
			"get": {
				"operationId": "SolanaAddMemo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Solana raw transaction with memo": {
										"value": "0a0b0c0d"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Solana",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tx",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "memo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/utils/solana/replace-fee-payer": {
			"get": {
				"operationId": "SponsorWrapRawTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Solana raw transaction with replaced fee payer": {
										"value": "0a0b0c0d"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Solana",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sponsor",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tx",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fundLamports",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/utils/stellar/submit": {
			"get": {
				"operationId": "SubmitTransactionStellar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HorizonApi.SubmitTransactionResponse"
								},
								"examples": {
									"Submit response": {
										"value": {
											"hash": "stellar-tx-hash"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Stellar"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "xdrTx",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/utils/soroban/send": {
			"get": {
				"operationId": "SendTransactionSoroban",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Api.SendTransactionResponse"
								},
								"examples": {
									"Send response": {
										"value": {
											"hash": "soroban-tx-hash",
											"status": "PENDING"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Soroban"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "xdrTx",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/utils/soroban/confirm": {
			"get": {
				"operationId": "ConfirmSorobanTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Api.GetTransactionResponse"
								},
								"examples": {
									"Confirmation response": {
										"value": {
											"status": "SUCCESS"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Soroban"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "hash",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "secondsToWait",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/utils/sui/build-send-from-custom-tx": {
			"get": {
				"operationId": "SuiBuildSendTxFromCustomTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Raw Sui bridge transaction built from custom tx": {
										"value": "{\"version\":2,\"sender\":\"0x0\",\"expiration\":{\"None\":true},\"gasData\":{\"budget\":\"1\",\"price\":\"1\",\"owner\":\"0x0\",\"payment\":[]}}"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"SUI",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "baseTx",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "inputCoin",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "params",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/utils/sui/raw2base64": {
			"get": {
				"operationId": "SuiRaw2Base64",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Base64 encoded Sui transaction": {
										"value": "AAECAw=="
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"SUI",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "rawTx",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/utils/tron/raw2hex": {
			"get": {
				"operationId": "TronRaw2Hex",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Hex encoded Tron transaction": {
										"value": "0a0b0c0d"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Utils",
					"Tron",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "rawTx",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}